Cpu: improved `Instruction`, removed `Register` and `PublicInstruction`
There is no need for weird wrapper types, we can use `RegisterType`
directly, and for better performance we can just use the `rd/rs/rt_raw`
values which are just numbers, exactly what `Register` was.

Also, removed `dead_code` usages in dma (forgot why it was put there),
and other places on the CPU.
