+++ psx-core/src/cpu.rs
@@ -1,15 +1,13 @@
 #[cfg(feature = "debugger")]
 mod debugger;
 mod instruction;
-mod instruction_format;
 mod instructions_table;
 mod register;
 
 use crate::coprocessor::{Gte, SystemControlCoprocessor};
 use crate::memory::BusLine;
 
-use instruction::Instruction;
-pub use instruction::{Opcode, PublicInstruction as CpuInstruction};
+pub use instruction::{Instruction, Opcode};
 pub use register::{RegisterType, Registers, CPU_REGISTERS};
 
 #[cfg(feature = "debugger")]
@@ -292,14 +290,14 @@ impl Cpu {
     where
         F: FnOnce(u32, u32) -> (u32, bool),
     {
-        let rs = self.regs.read_general(instruction.rs);
-        let rt = self.regs.read_general(instruction.rt);
+        let rs = self.regs.read_general(instruction.rs_raw);
+        let rt = self.regs.read_general(instruction.rt_raw);
 
         let (result, overflow) = handler(rs, rt);
         if overflow {
             self.execute_exception(Exception::ArithmeticOverflow);
         } else {
-            self.regs.write_general(instruction.rd, result);
+            self.regs.write_general(instruction.rd_raw, result);
         }
     }
 
@@ -308,9 +306,9 @@ impl Cpu {
     where
         F: FnOnce(u32, &Instruction) -> u32,
     {
-        let rs = self.regs.read_general(instruction.rs);
+        let rs = self.regs.read_general(instruction.rs_raw);
         let result = handler(rs, &instruction);
-        self.regs.write_general(instruction.rt, result);
+        self.regs.write_general(instruction.rt_raw, result);
     }
 
     #[inline]
@@ -318,9 +316,9 @@ impl Cpu {
     where
         F: FnOnce(i32, i32) -> bool,
     {
-        let rs = self.regs.read_general(instruction.rs) as i32;
+        let rs = self.regs.read_general(instruction.rs_raw) as i32;
         let rt = if have_rt {
-            self.regs.read_general(instruction.rt) as i32
+            self.regs.read_general(instruction.rt_raw) as i32
         } else {
             0
         };
@@ -337,11 +335,11 @@ impl Cpu {
     where
         F: FnMut(&mut Self, u32) -> Option<u32>,
     {
-        let rs = self.regs.read_general(instruction.rs);
+        let rs = self.regs.read_general(instruction.rs_raw);
         let computed_addr = rs.wrapping_add(Self::sign_extend_16(instruction.imm16));
 
         if let Some(data) = handler(self, computed_addr) {
-            self.regs.write_general(instruction.rt, data);
+            self.regs.write_general(instruction.rt_raw, data);
         }
     }
 
@@ -350,8 +348,8 @@ impl Cpu {
     where
         F: FnMut(&mut Self, u32, u32),
     {
-        let rs = self.regs.read_general(instruction.rs);
-        let rt = self.regs.read_general(instruction.rt);
+        let rs = self.regs.read_general(instruction.rs_raw);
+        let rt = self.regs.read_general(instruction.rt_raw);
         let computed_addr = rs.wrapping_add(Self::sign_extend_16(instruction.imm16));
 
         handler(self, computed_addr, rt);
@@ -389,7 +387,7 @@ impl Cpu {
             }
             Opcode::Lwl => {
                 // TODO: test these unaligned addressing instructions
-                let rs = self.regs.read_general(instruction.rs);
+                let rs = self.regs.read_general(instruction.rs_raw);
                 let computed_addr = rs.wrapping_add(Self::sign_extend_16(instruction.imm16));
 
                 // round to the nearest floor of four
@@ -408,13 +406,13 @@ impl Cpu {
                 result <<= shift;
 
                 let mask = !((0xFFFFFFFF >> shift) << shift);
-                let original_rt = self.regs.read_general(instruction.rt);
+                let original_rt = self.regs.read_general(instruction.rt_raw);
                 let result = (original_rt & mask) | result;
 
-                self.regs.write_general(instruction.rt, result);
+                self.regs.write_general(instruction.rt_raw, result);
             }
             Opcode::Lwr => {
-                let rs = self.regs.read_general(instruction.rs);
+                let rs = self.regs.read_general(instruction.rs_raw);
                 let computed_addr = rs.wrapping_add(Self::sign_extend_16(instruction.imm16));
 
                 let start = computed_addr;
@@ -431,10 +429,10 @@ impl Cpu {
                 let shift = offset * 8;
 
                 let mask = !(0xFFFFFFFF >> shift);
-                let original_rt = self.regs.read_general(instruction.rt);
+                let original_rt = self.regs.read_general(instruction.rt_raw);
                 let result = (original_rt & mask) | result;
 
-                self.regs.write_general(instruction.rt, result);
+                self.regs.write_general(instruction.rt_raw, result);
             }
             Opcode::Sb => {
                 self.execute_store(instruction, |s, computed_addr, data| {
@@ -452,8 +450,8 @@ impl Cpu {
                 });
             }
             Opcode::Swl => {
-                let rs = self.regs.read_general(instruction.rs);
-                let mut rt = self.regs.read_general(instruction.rt);
+                let rs = self.regs.read_general(instruction.rs_raw);
+                let mut rt = self.regs.read_general(instruction.rt_raw);
                 let computed_addr = rs.wrapping_add(Self::sign_extend_16(instruction.imm16));
 
                 // round to the nearest floor of four
@@ -472,8 +470,8 @@ impl Cpu {
                 }
             }
             Opcode::Swr => {
-                let rs = self.regs.read_general(instruction.rs);
-                let mut rt = self.regs.read_general(instruction.rt);
+                let rs = self.regs.read_general(instruction.rs_raw);
+                let mut rt = self.regs.read_general(instruction.rt_raw);
                 let computed_addr = rs.wrapping_add(Self::sign_extend_16(instruction.imm16));
 
                 let start = computed_addr;
@@ -486,28 +484,32 @@ impl Cpu {
                 }
             }
             Opcode::Slt => {
-                let rs = self.regs.read_general(instruction.rs) as i32;
-                let rt = self.regs.read_general(instruction.rt) as i32;
+                let rs = self.regs.read_general(instruction.rs_raw) as i32;
+                let rt = self.regs.read_general(instruction.rt_raw) as i32;
 
-                self.regs.write_general(instruction.rd, (rs < rt) as u32);
+                self.regs
+                    .write_general(instruction.rd_raw, (rs < rt) as u32);
             }
             Opcode::Sltu => {
-                let rs = self.regs.read_general(instruction.rs);
-                let rt = self.regs.read_general(instruction.rt);
+                let rs = self.regs.read_general(instruction.rs_raw);
+                let rt = self.regs.read_general(instruction.rt_raw);
 
-                self.regs.write_general(instruction.rd, (rs < rt) as u32);
+                self.regs
+                    .write_general(instruction.rd_raw, (rs < rt) as u32);
             }
             Opcode::Slti => {
-                let rs = self.regs.read_general(instruction.rs) as i32;
+                let rs = self.regs.read_general(instruction.rs_raw) as i32;
                 let imm = instruction.imm16 as i16 as i32;
 
-                self.regs.write_general(instruction.rt, (rs < imm) as u32);
+                self.regs
+                    .write_general(instruction.rt_raw, (rs < imm) as u32);
             }
             Opcode::Sltiu => {
-                let rs = self.regs.read_general(instruction.rs);
+                let rs = self.regs.read_general(instruction.rs_raw);
                 let imm = Self::sign_extend_16(instruction.imm16);
 
-                self.regs.write_general(instruction.rt, (rs < imm) as u32);
+                self.regs
+                    .write_general(instruction.rt_raw, (rs < imm) as u32);
             }
             Opcode::Addu => {
                 self.execute_alu_reg(instruction, |rs, rt| (rs.wrapping_add(rt), false));
@@ -533,14 +535,14 @@ impl Cpu {
                 });
             }
             Opcode::Addi => {
-                let rs = self.regs.read_general(instruction.rs);
+                let rs = self.regs.read_general(instruction.rs_raw);
                 let (result, overflow) =
                     (rs as i32).overflowing_add(Self::sign_extend_16(instruction.imm16) as i32);
 
                 if overflow {
                     self.execute_exception(Exception::ArithmeticOverflow);
                 } else {
-                    self.regs.write_general(instruction.rt, result as u32);
+                    self.regs.write_general(instruction.rt_raw, result as u32);
                 }
             }
             Opcode::And => {
@@ -576,28 +578,28 @@ impl Cpu {
                 });
             }
             Opcode::Sll => {
-                let rt = self.regs.read_general(instruction.rt);
+                let rt = self.regs.read_general(instruction.rt_raw);
                 let result = rt << instruction.imm5;
-                self.regs.write_general(instruction.rd, result);
+                self.regs.write_general(instruction.rd_raw, result);
             }
             Opcode::Srl => {
-                let rt = self.regs.read_general(instruction.rt);
+                let rt = self.regs.read_general(instruction.rt_raw);
                 let result = rt >> instruction.imm5;
-                self.regs.write_general(instruction.rd, result);
+                self.regs.write_general(instruction.rd_raw, result);
             }
             Opcode::Sra => {
-                let rt = self.regs.read_general(instruction.rt);
+                let rt = self.regs.read_general(instruction.rt_raw);
                 let result = ((rt as i32) >> instruction.imm5) as u32;
-                self.regs.write_general(instruction.rd, result);
+                self.regs.write_general(instruction.rd_raw, result);
             }
             Opcode::Lui => {
                 let result = (instruction.imm16 as u32) << 16;
-                self.regs.write_general(instruction.rt, result);
+                self.regs.write_general(instruction.rt_raw, result);
             }
             Opcode::Mult => {
                 self.elapsed_cycles += 5;
-                let rs = self.regs.read_general(instruction.rs) as i32 as i64;
-                let rt = self.regs.read_general(instruction.rt) as i32 as i64;
+                let rs = self.regs.read_general(instruction.rs_raw) as i32 as i64;
+                let rt = self.regs.read_general(instruction.rt_raw) as i32 as i64;
 
                 let result = (rs * rt) as u64;
 
@@ -606,8 +608,8 @@ impl Cpu {
             }
             Opcode::Multu => {
                 self.elapsed_cycles += 5;
-                let rs = self.regs.read_general(instruction.rs) as u64;
-                let rt = self.regs.read_general(instruction.rt) as u64;
+                let rs = self.regs.read_general(instruction.rs_raw) as u64;
+                let rt = self.regs.read_general(instruction.rt_raw) as u64;
 
                 let result = rs * rt;
 
@@ -616,8 +618,8 @@ impl Cpu {
             }
             Opcode::Div => {
                 self.elapsed_cycles += 10;
-                let rs = self.regs.read_general(instruction.rs) as i32 as i64;
-                let rt = self.regs.read_general(instruction.rt) as i32 as i64;
+                let rs = self.regs.read_general(instruction.rs_raw) as i32 as i64;
+                let rt = self.regs.read_general(instruction.rt_raw) as i32 as i64;
 
                 // division by zero (overflow)
                 if rt == 0 {
@@ -634,8 +636,8 @@ impl Cpu {
             }
             Opcode::Divu => {
                 self.elapsed_cycles += 10;
-                let rs = self.regs.read_general(instruction.rs) as u64;
-                let rt = self.regs.read_general(instruction.rt) as u64;
+                let rs = self.regs.read_general(instruction.rs_raw) as u64;
+                let rt = self.regs.read_general(instruction.rt_raw) as u64;
 
                 // division by zero
                 if rt == 0 {
@@ -650,16 +652,16 @@ impl Cpu {
                 }
             }
             Opcode::Mfhi => {
-                self.regs.write_general(instruction.rd, self.regs.hi);
+                self.regs.write_general(instruction.rd_raw, self.regs.hi);
             }
             Opcode::Mthi => {
-                self.regs.hi = self.regs.read_general(instruction.rs);
+                self.regs.hi = self.regs.read_general(instruction.rs_raw);
             }
             Opcode::Mflo => {
-                self.regs.write_general(instruction.rd, self.regs.lo);
+                self.regs.write_general(instruction.rd_raw, self.regs.lo);
             }
             Opcode::Mtlo => {
-                self.regs.lo = self.regs.read_general(instruction.rs);
+                self.regs.lo = self.regs.read_general(instruction.rs_raw);
             }
             Opcode::J => {
                 let base = self.regs.pc & 0xF0000000;
@@ -676,12 +678,13 @@ impl Cpu {
                 self.regs.write_ra(self.regs.pc + 4);
             }
             Opcode::Jr => {
-                self.jump_dest_next = Some(self.regs.read_general(instruction.rs));
+                self.jump_dest_next = Some(self.regs.read_general(instruction.rs_raw));
             }
             Opcode::Jalr => {
-                self.jump_dest_next = Some(self.regs.read_general(instruction.rs));
+                self.jump_dest_next = Some(self.regs.read_general(instruction.rs_raw));
 
-                self.regs.write_general(instruction.rd, self.regs.pc + 4);
+                self.regs
+                    .write_general(instruction.rd_raw, self.regs.pc + 4);
             }
             Opcode::Beq => {
                 self.execute_branch(instruction, true, |rs, rt| rs == rt);
@@ -732,7 +735,7 @@ impl Cpu {
                     _ => unreachable!(),
                 };
 
-                self.regs.write_general(instruction.rt, result);
+                self.regs.write_general(instruction.rt_raw, result);
             }
             Opcode::Cfc(n) => {
                 let result = match n {
@@ -741,10 +744,10 @@ impl Cpu {
                     _ => unreachable!(),
                 };
 
-                self.regs.write_general(instruction.rt, result);
+                self.regs.write_general(instruction.rt_raw, result);
             }
             Opcode::Mtc(n) => {
-                let rt = self.regs.read_general(instruction.rt);
+                let rt = self.regs.read_general(instruction.rt_raw);
 
                 match n {
                     0 => self.cop0.write_data(instruction.rd_raw, rt),
@@ -753,7 +756,7 @@ impl Cpu {
                 }
             }
             Opcode::Ctc(n) => {
-                let rt = self.regs.read_general(instruction.rt);
+                let rt = self.regs.read_general(instruction.rt_raw);
 
                 match n {
                     0 => self.cop0.write_ctrl(instruction.rd_raw, rt),
@@ -775,7 +778,7 @@ impl Cpu {
                 self.cop0.write_sr(sr);
             }
             Opcode::Lwc(n) => {
-                let rs = self.regs.read_general(instruction.rs);
+                let rs = self.regs.read_general(instruction.rs_raw);
                 let computed_addr = rs.wrapping_add(Self::sign_extend_16(instruction.imm16));
 
                 if let Some(data) = self.bus_read_u32(bus, computed_addr) {
