+++ psx-core/src/cpu/instruction.rs
@@ -1,5 +1,6 @@
+use std::fmt;
+
 use super::instructions_table::{PRIMARY_OPCODES, SECONDARY_OPCODES};
-use super::register::Register;
 use super::RegisterType;
 
 #[derive(Copy, Clone, Debug)]
@@ -111,6 +112,15 @@ pub enum Opcode {
     Swc(u8),
 }
 
+/// MIPS Instruction representation.
+///
+/// Do note that the fields may not always make sense as there are fields used
+/// by some instructions and some not used.
+///
+/// For example, the instruction `Lw` may have the value for `rd` set to
+/// `RegisterType::A1`, or `RegisterType::T1` and it shouldn't matter
+/// since that instruction only uses `rs` and `rt`. So keep that in mind when
+/// accessing the fields.
 #[derive(Debug, Clone)]
 pub struct Instruction {
     pub pc: u32,
@@ -119,11 +129,11 @@ pub struct Instruction {
 
     pub imm5: u8,
     pub rd_raw: u8,
-    pub rd: Register,
+    pub rd: RegisterType,
     pub rt_raw: u8,
-    pub rt: Register,
+    pub rt: RegisterType,
     pub rs_raw: u8,
-    pub rs: Register,
+    pub rs: RegisterType,
     pub imm16: u16,
     pub imm25: u32,
     pub imm26: u32,
@@ -138,11 +148,11 @@ impl Instruction {
                 opcode: Opcode::Nop,
                 imm5: 0,
                 rd_raw: 0,
-                rd: Register::from_byte(0),
+                rd: RegisterType::from(0),
                 rt_raw: 0,
-                rt: Register::from_byte(0),
+                rt: RegisterType::from(0),
                 rs_raw: 0,
-                rs: Register::from_byte(0),
+                rs: RegisterType::from(0),
                 imm16: 0,
                 imm25: 0,
                 imm26: 0,
@@ -153,11 +163,11 @@ impl Instruction {
         let secondary_identifier = instruction as u8 & 0x3F;
         let imm5 = (instruction >> 6) as u8 & 0x1F;
         let rd_raw = (instruction >> 11) as u8 & 0x1F;
-        let rd = Register::from_byte(rd_raw);
+        let rd = RegisterType::from(rd_raw);
         let rt_raw = (instruction >> 16) as u8 & 0x1F;
-        let rt = Register::from_byte(rt_raw);
+        let rt = RegisterType::from(rt_raw);
         let rs_raw = (instruction >> 21) as u8 & 0x1F;
-        let rs = Register::from_byte(rs_raw);
+        let rs = RegisterType::from(rs_raw);
         // combination of the above
         let imm16 = instruction as u16;
         let imm26 = instruction & 0x3FFFFFF;
@@ -189,7 +199,6 @@ impl Instruction {
         }
     }
 
-    #[allow(dead_code)]
     pub fn is_branch(&self) -> bool {
         match self.opcode {
             Opcode::J
@@ -257,51 +266,273 @@ impl Instruction {
     }
 }
 
-// A struct that uses `RegisterType`, which is good for public access but less
-// optimial to generate. That's why we are using this ugly solution of wrapping our own
-// `Instruction`, and this resulted in not being able to really find a good name for it.
-/// Instruction representation.
-#[derive(Clone)]
-pub struct PublicInstruction {
-    pub opcode: Opcode,
-
-    pub imm5: u8,
-    pub rd: RegisterType,
-    pub rt: RegisterType,
-    pub rs: RegisterType,
-    pub imm16: u16,
-    pub imm25: u32,
-    pub imm26: u32,
-
-    inner: Instruction,
+const fn opcode_str(opcode: Opcode) -> &'static str {
+    match opcode {
+        Opcode::Nop => "nop",
+        Opcode::Lb => "lb",
+        Opcode::Lh => "lh",
+        Opcode::Lw => "lw",
+        Opcode::Lbu => "lbu",
+        Opcode::Lhu => "lhu",
+        Opcode::Sb => "sb",
+        Opcode::Lwl => "lwl",
+        Opcode::Lwr => "lwr",
+        Opcode::Sh => "sh",
+        Opcode::Sw => "sw",
+        Opcode::Swl => "swl",
+        Opcode::Swr => "swr",
+        Opcode::Slt => "slt",
+        Opcode::Sltu => "sltu",
+        Opcode::Slti => "slti",
+        Opcode::Sltiu => "sltiu",
+        Opcode::Addu => "addu",
+        Opcode::Add => "add",
+        Opcode::Subu => "subu",
+        Opcode::Sub => "sub",
+        Opcode::Addiu => "addiu",
+        Opcode::Addi => "addi",
+        Opcode::And => "and",
+        Opcode::Or => "or",
+        Opcode::Xor => "xor",
+        Opcode::Nor => "nor",
+        Opcode::Andi => "andi",
+        Opcode::Ori => "ori",
+        Opcode::Xori => "xori",
+        Opcode::Sllv => "sllv",
+        Opcode::Srlv => "srlv",
+        Opcode::Srav => "srav",
+        Opcode::Sll => "sll",
+        Opcode::Srl => "srl",
+        Opcode::Sra => "sra",
+        Opcode::Lui => "lui",
+        Opcode::Mult => "mult",
+        Opcode::Multu => "multu",
+        Opcode::Div => "div",
+        Opcode::Divu => "divu",
+        Opcode::Mfhi => "mfhi",
+        Opcode::Mthi => "mthi",
+        Opcode::Mflo => "mflo",
+        Opcode::Mtlo => "mtlo",
+        Opcode::J => "j",
+        Opcode::Jal => "jal",
+        Opcode::Jr => "jr",
+        Opcode::Jalr => "jalr",
+        Opcode::Beq => "beq",
+        Opcode::Bne => "bne",
+        Opcode::Bgtz => "bgtz",
+        Opcode::Blez => "blez",
+        Opcode::Bcondz => "bcondz",
+        Opcode::Bltz => "bltz",
+        Opcode::Bgez => "bgez",
+        Opcode::Bltzal => "bltzal",
+        Opcode::Bgezal => "bgezal",
+        Opcode::Syscall => "syscall",
+        Opcode::Break => "break",
+        Opcode::Cop(0) => "cop0",
+        Opcode::Cop(1) => "cop1",
+        Opcode::Cop(2) => "cop2",
+        Opcode::Cop(3) => "cop3",
+        Opcode::Mfc(0) => "mfc0",
+        Opcode::Mfc(1) => "mfc1",
+        Opcode::Mfc(2) => "mfc2",
+        Opcode::Mfc(3) => "mfc3",
+        Opcode::Cfc(0) => "cfc0",
+        Opcode::Cfc(1) => "cfc1",
+        Opcode::Cfc(2) => "cfc2",
+        Opcode::Cfc(3) => "cfc3",
+        Opcode::Mtc(0) => "mtc0",
+        Opcode::Mtc(1) => "mtc1",
+        Opcode::Mtc(2) => "mtc2",
+        Opcode::Mtc(3) => "mtc3",
+        Opcode::Ctc(0) => "ctc0",
+        Opcode::Ctc(1) => "ctc1",
+        Opcode::Ctc(2) => "ctc2",
+        Opcode::Ctc(3) => "ctc3",
+        Opcode::Bcf(0) => "bcf0",
+        Opcode::Bcf(1) => "bcf1",
+        Opcode::Bcf(2) => "bcf2",
+        Opcode::Bcf(3) => "bcf3",
+        Opcode::Bct(0) => "bct0",
+        Opcode::Bct(1) => "bct1",
+        Opcode::Bct(2) => "bct2",
+        Opcode::Bct(3) => "bct3",
+        Opcode::Rfe => "rfe",
+        Opcode::Lwc(0) => "lwc0",
+        Opcode::Lwc(1) => "lwc1",
+        Opcode::Lwc(2) => "lwc2",
+        Opcode::Lwc(3) => "lwc3",
+        Opcode::Swc(0) => "swc0",
+        Opcode::Swc(1) => "swc1",
+        Opcode::Swc(2) => "swc2",
+        Opcode::Swc(3) => "swc3",
+        _ => unreachable!(),
+    }
 }
 
-impl PublicInstruction {
-    pub fn from_u32(data: u32) -> Self {
-        let inner = Instruction::from_u32(data, 0);
+fn format_load_store(f: &mut fmt::Formatter, instr: &Instruction) -> fmt::Result {
+    let opcode = instr.opcode;
 
-        Self {
-            opcode: inner.opcode,
+    let src = instr.rs;
+    let off = instr.imm16;
+    let dst = instr.rt;
 
-            imm5: inner.imm5,
-            rd: inner.rd.into(),
-            rt: inner.rt.into(),
-            rs: inner.rs.into(),
-            imm16: inner.imm16,
-            imm25: inner.imm25,
-            imm26: inner.imm26,
+    write!(f, "{} {}, 0x{:04X}({})", opcode_str(opcode), dst, off, src)
+}
 
-            inner,
+fn format_alu(f: &mut fmt::Formatter, instr: &Instruction, imm: bool) -> fmt::Result {
+    let opcode = instr.opcode;
+
+    let (first, third) = if imm {
+        (instr.rt, format!("0x{:04X}", instr.imm16))
+    } else {
+        (instr.rd, format!("{}", instr.rt))
+    };
+
+    write!(
+        f,
+        "{} {}, {}, {}",
+        opcode_str(opcode),
+        first,
+        instr.rs,
+        third
+    )
+}
+
+fn format_shift(f: &mut fmt::Formatter, instr: &Instruction, imm: bool) -> fmt::Result {
+    let opcode = instr.opcode;
+
+    let third = if imm {
+        format!("0x{:02X}", instr.imm5)
+    } else {
+        format!("{}", instr.rs)
+    };
+
+    write!(
+        f,
+        "{} {}, {}, {}",
+        opcode_str(opcode),
+        instr.rd,
+        instr.rt,
+        third
+    )
+}
+
+fn format_mult_div(f: &mut fmt::Formatter, instr: &Instruction) -> fmt::Result {
+    let opcode = instr.opcode;
+
+    write!(f, "{} {}, {}", opcode_str(opcode), instr.rs, instr.rt)
+}
+
+fn format_branch(f: &mut fmt::Formatter, instr: &Instruction, rt: bool) -> fmt::Result {
+    let opcode = instr.opcode;
+
+    let dest = instr.imm16;
+
+    if rt {
+        write!(
+            f,
+            "{} {}, {}, 0x{:04X} => 0x{:08X}",
+            opcode_str(opcode),
+            instr.rs,
+            instr.rt,
+            dest,
+            instr
+                .pc
+                .wrapping_add((dest as i16 as i32 as u32).wrapping_mul(4))
+                .wrapping_add(4)
+        )
+    } else {
+        write!(
+            f,
+            "{} {}, 0x{:04X} => 0x{:08X}",
+            opcode_str(opcode),
+            instr.rs,
+            dest,
+            instr
+                .pc
+                .wrapping_add((dest as i16 as i32 as u32).wrapping_mul(4))
+                .wrapping_add(4)
+        )
+    }
+}
+
+fn format_cop_ops(f: &mut fmt::Formatter, instr: &Instruction) -> fmt::Result {
+    let opcode = instr.opcode;
+
+    write!(f, "{} {}, {}", opcode_str(opcode), instr.rt, instr.rd)
+}
+
+impl fmt::Display for Instruction {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self.opcode {
+            Opcode::Lb
+            | Opcode::Lbu
+            | Opcode::Lh
+            | Opcode::Lhu
+            | Opcode::Lw
+            | Opcode::Lwl
+            | Opcode::Lwr
+            | Opcode::Sb
+            | Opcode::Sh
+            | Opcode::Sw
+            | Opcode::Swl
+            | Opcode::Swr => format_load_store(f, self),
+            Opcode::Slt | Opcode::Sltu => format_alu(f, self, false),
+            Opcode::Slti | Opcode::Sltiu => format_alu(f, self, true),
+            Opcode::Addu | Opcode::Add | Opcode::Subu | Opcode::Sub => format_alu(f, self, false),
+            Opcode::Addiu | Opcode::Addi => format_alu(f, self, true),
+            Opcode::And | Opcode::Or | Opcode::Xor | Opcode::Nor => format_alu(f, self, false),
+            Opcode::Andi | Opcode::Ori | Opcode::Xori => format_alu(f, self, true),
+            Opcode::Sllv | Opcode::Srlv | Opcode::Srav => format_shift(f, self, false),
+            Opcode::Sll | Opcode::Srl | Opcode::Sra => format_shift(f, self, true),
+            Opcode::Lui => write!(
+                f,
+                "{} {}, 0x{:04X}",
+                opcode_str(self.opcode),
+                self.rt,
+                self.imm16
+            ),
+            Opcode::Mult | Opcode::Multu | Opcode::Div | Opcode::Divu => format_mult_div(f, self),
+            Opcode::Mfhi => write!(f, "{} {}", opcode_str(self.opcode), self.rd),
+            Opcode::Mthi => write!(f, "{} {}", opcode_str(self.opcode), self.rs),
+            Opcode::Mflo => write!(f, "{} {}", opcode_str(self.opcode), self.rd),
+            Opcode::Mtlo => write!(f, "{} {}", opcode_str(self.opcode), self.rs),
+            Opcode::J => write!(
+                f,
+                "{} 0x{:07X} => 0x{:08X}",
+                opcode_str(self.opcode),
+                self.imm26,
+                self.pc & 0xF0000000 | self.imm26 * 4
+            ),
+            Opcode::Jal => write!(
+                f,
+                "{} 0x{:07X} => 0x{:08X}",
+                opcode_str(self.opcode),
+                self.imm26,
+                self.pc & 0xF0000000 | self.imm26 * 4
+            ),
+            Opcode::Jr => write!(f, "{} {}", opcode_str(self.opcode), self.rs),
+            // some specs says "jalr rd,rs"
+            Opcode::Jalr => write!(f, "{} {}, {}", opcode_str(self.opcode), self.rs, self.rd),
+            Opcode::Beq | Opcode::Bne => format_branch(f, self, true),
+            Opcode::Bgtz
+            | Opcode::Blez
+            | Opcode::Bltz
+            | Opcode::Bgez
+            | Opcode::Bltzal
+            | Opcode::Bgezal => format_branch(f, self, false),
+            Opcode::Nop | Opcode::Syscall | Opcode::Break | Opcode::Rfe => {
+                f.write_str(opcode_str(self.opcode))
+            }
+            Opcode::Cop(_) => write!(f, "{} 0x{:07X}", opcode_str(self.opcode), self.imm25),
+            Opcode::Mfc(_) | Opcode::Cfc(_) | Opcode::Mtc(_) | Opcode::Ctc(_) => {
+                format_cop_ops(f, self)
+            }
+            Opcode::Bcf(_) => todo!(),
+            Opcode::Bct(_) => todo!(),
+            Opcode::Lwc(_) | Opcode::Swc(_) => format_load_store(f, self),
+            Opcode::Invalid => write!(f, "Invalid instruction"),
+            _ => unreachable!(),
         }
-    }
-
-    pub fn is_branch(&self) -> bool {
-        self.inner.is_branch()
-    }
-}
-
-impl std::fmt::Display for PublicInstruction {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.inner.fmt(f)
     }
 }
